+++ controller/konnect/ops.go
@@ -13,17 +13,25 @@ import (
 
 	"github.com/kong/gateway-operator/controller/pkg/log"
 	k8sutils "github.com/kong/gateway-operator/pkg/utils/kubernetes"
+
+	konnectv1alpha1 "github.com/kong/kubernetes-configuration/api/konnect/v1alpha1"
 )
 
-// Op is the Konnect operation type.
+// Response is the interface for the response from the Konnect API.
+type Response interface {
+	GetContentType() string
+	GetStatusCode() int
+}
+
+// Op is the type for the operation type of a Konnect entity.
 type Op string
 
 const (
-	// CreateOp is the Konnect create operation.
+	// CreateOp is the operation type for creating a Konnect entity.
 	CreateOp Op = "create"
-	// UpdateOp is the Konnect update operation.
+	// UpdateOp is the operation type for updating a Konnect entity.
 	UpdateOp Op = "update"
-	// DeleteOp is the Konnect delete operation.
+	// DeleteOp is the operation type for deleting a Konnect entity.
 	DeleteOp Op = "delete"
 )
 
@@ -34,9 +42,12 @@ func Create[
 ](ctx context.Context, sdk *sdkkonnectgo.SDK, logger logr.Logger, cl client.Client, e *T) (*T, error) {
 	defer logOpComplete[T, TEnt](logger, time.Now(), CreateOp, e)
 
-	switch ent := any(e).(type) { //nolint:gocritic
-	// ---------------------------------------------------------------------
-	// TODO: add other Konnect types
+	switch ent := any(e).(type) {
+	case *konnectv1alpha1.KonnectControlPlane:
+		return e, createControlPlane(ctx, sdk, logger, ent)
+
+		// ---------------------------------------------------------------------
+		// TODO: add other Konnect types
 
 	default:
 		return nil, fmt.Errorf("unsupported entity type %T", ent)
@@ -50,9 +61,12 @@ func Delete[
 ](ctx context.Context, sdk *sdkkonnectgo.SDK, logger logr.Logger, cl client.Client, e *T) error {
 	defer logOpComplete[T, TEnt](logger, time.Now(), DeleteOp, e)
 
-	switch ent := any(e).(type) { //nolint:gocritic
-	// ---------------------------------------------------------------------
-	// TODO: add other Konnect types
+	switch ent := any(e).(type) {
+	case *konnectv1alpha1.KonnectControlPlane:
+		return deleteControlPlane(ctx, sdk, logger, ent)
+
+		// ---------------------------------------------------------------------
+		// TODO: add other Konnect types
 
 	default:
 		return fmt.Errorf("unsupported entity type %T", ent)
@@ -91,9 +105,12 @@ func Update[
 
 	defer logOpComplete[T, TEnt](logger, now, UpdateOp, e)
 
-	switch ent := any(e).(type) { //nolint:gocritic
-	// ---------------------------------------------------------------------
-	// TODO: add other Konnect types
+	switch ent := any(e).(type) {
+	case *konnectv1alpha1.KonnectControlPlane:
+		return ctrl.Result{}, updateControlPlane(ctx, sdk, logger, ent)
+
+		// ---------------------------------------------------------------------
+		// TODO: add other Konnect types
 
 	default:
 		return ctrl.Result{}, fmt.Errorf("unsupported entity type %T", ent)
@@ -110,4 +127,19 @@ func logOpComplete[
 		"type", entityTypeName[T](),
 		"konnect_id", e.GetKonnectStatus().GetKonnectID(),
 	)
+}
+
+// wrapErrIfKonnectOpFailed checks the response from the Konnect API and returns a uniform
+// error for all Konnect entities if the operation failed.
+func wrapErrIfKonnectOpFailed[
+	T SupportedKonnectEntityType,
+	TEnt EntityType[T],
+](err error, op Op) error {
+	if err != nil {
+		var e TEnt
+		return fmt.Errorf("failed to %s for %T %q: %w",
+			op, client.ObjectKeyFromObject(e), e, err,
+		)
+	}
+	return nil
 }